Improve `K` behaviour
I reviewed most patterns for completeness and added a new heuristic that
"parses" the import lines for easier usage. When you use aliased
modules or functions in the code, it should be able to pickup what you
really meant.
It's still just regex trickery, it's not a real parser, so it might fail
on edge cases.
Should be pretty robust, here's some case it detects. `a`, `b` and `c`
can be submodules (having internal dots).

```python
from a import b as c
from a import b
import b as c
```

Thanks to @comeonfox for including a PR upstream that made me review
this.
https://github.com/fs111/pydoc.vim/pull/34